Move MousePosition code into Editor
