+++ S2VX.Game/Story.cs
@@ -26,8 +26,6 @@ namespace S2VX.Game
         public double GameTime { get; private set; } = 0;
         public bool IsPlaying { get; private set; } = false;
 
-        public Vector2 MousePosition { get; private set; } = Vector2.Zero;
-
         public Camera Camera { get; } = new Camera();
         public RelativeBox Background = new RelativeBox
         {
@@ -140,13 +138,6 @@ namespace S2VX.Game
             };
             var serialized = JsonConvert.SerializeObject(obj, Formatting.Indented, converters);
             File.WriteAllText(path, serialized);
-        }
-
-        protected override bool OnMouseMove(MouseMoveEvent e)
-        {
-
-            MousePosition = ToLocalSpace(e.CurrentState.Mouse.Position);
-            return true;
         }
 
         protected override void Update()
